application/backend/pyproject.toml
[project]
name = "geti-action"
version = "1.0.0"
description = "Geti Action server"
requires-python = ">=3.12"

dependencies = [
  "opencv-python",
  "cv2-enumerate-cameras",
  'harvesters; sys_platform != "darwin"',
  "fastapi[standard]",
  "lerobot[feetech]==0.4.2",
  "numpy",
  "websockets",
  'pyrealsense2-macosx; sys_platform == "darwin"',
  'pyrealsense2; sys_platform != "darwin"',
  "aiortc>=1.13.0",
  "pydantic-settings>=2.10.1",
  "sqlalchemy>=2.0.43",
  "aiosqlite~=0.21",
  "alembic>=1.16.5",
  "click>=8.3.0",
  "FrameSource>=0.2.2",
  "loguru>=0.7.3",
  "getiaction[pi0,smolvla]",
  "greenlet>=3.2.4",
  "tenacity>=9.1.2",
  "aiofiles>=25.1.0",
  "types-aiofiles>=25.1.0.20251011",
  "trossen-arm==1.9.0"
]

[tool.uv.sources]
FrameSource = { git = "https://github.com/ArendJanKramer/FrameSource.git", rev = "4d51da9bc2f8f9353562e1641ddbf87d3045baf6" }
getiaction = { path = "../../library", editable = true }

[dependency-groups]
dev = [
  "ruff~=0.11.2",
]
lint = [
  "ruff~=0.11.2",
  "mypy~=1.17",
]
all = [
  "ruff~=0.11.2",
  "mypy~=1.17",
]

[tool.uv]
default-groups = ["all"]


# ============================================================================ #
# RUFF CONFIGURATION                                                           #
# ============================================================================ #
[tool.ruff]
# Extend shared configuration from root
extend = "../../pyproject.toml"

# Disable preview mode for backend (inherited as true from root)
preview = false

# Backend-specific source directories
src = ["src"]

# Backend-specific excludes
extend-exclude = [".venv*"]

[tool.ruff.lint]
# Backend uses a subset of rules (different from library's ALL)
select = [
    "ARG", "E", "F", "I", "N", "UP", "YTT", "ASYNC", "S", "COM", "C4", "FA",
    "PIE", "PYI", "Q", "RSE", "RET", "SIM", "TID", "TC", "PL", "RUF", "C90",
    "D103", "ANN001", "ANN201", "ANN205", "FAST"
]

ignore = [
    "N801", "N805", "N806", "N807", "N818", "COM812", "RET503", "SIM108",
    "SIM105", "PLR2004", "RUF010", "TC001", "RUF012"
]

fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"*test*.py" = ["S", "ANN", "D", "ARG", "PLR"]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pylint]
max-args = 7


# ============================================================================ #
# MYPY CONFIGURATION                                                           #
# ============================================================================ #
[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
show_error_codes = true
